# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yoriai API",
    "version": "0.1.0",
    "description": "yoriai バックエンド（Hono + Prisma + Supabase Auth）のAPI仕様書。認証は Supabase のセッションクッキーで行う。"
  },
  "servers": [
    {
      "url": "/",
      "description": "デフォルト"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "supabaseSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase Auth のセッションクッキーによる認証。実際のクッキー名はSupabaseプロジェクトごとに異なる（例: sb-<project-ref>-auth-token）。"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "QuestionPagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "ModerationResult": {
        "type": "object",
        "properties": {
          "isHidden": {
            "type": "boolean",
            "description": "AI判定によって非公開になった場合はtrue"
          }
        },
        "required": [
          "isHidden"
        ]
      },
      "TagAssignmentStatus": {
        "type": "string",
        "enum": [
          "assigned",
          "failed",
          "skipped"
        ]
      },
      "TagAssignmentErrorStatus": {
        "anyOf": [
          {
            "type": "number",
            "enum": [
              422
            ]
          },
          {
            "type": "number",
            "enum": [
              429
            ]
          },
          {
            "type": "number",
            "enum": [
              502
            ]
          },
          {
            "type": "number",
            "enum": [
              503
            ]
          },
          {
            "type": "number",
            "enum": [
              504
            ]
          }
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/questions": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "質問一覧を検索・絞り込み・ページングして取得",
        "security": [
          {
            "supabaseSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "required": false,
            "name": "keyword",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "unanswered",
                "resolved"
              ],
              "default": "all"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "tagId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "tag-category-1,tag-category-2"
            },
            "required": false,
            "name": "categoryIds",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "tag-1,tag-2"
            },
            "required": false,
            "name": "tagIds",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "質問一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "questions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Question"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/QuestionPagination"
                    }
                  },
                  "required": [
                    "questions",
                    "pagination"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "質問を作成",
        "security": [
          {
            "supabaseSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 48
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tagId": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "title",
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "同一内容の再送",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "question": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "moderation": {
                      "$ref": "#/components/schemas/ModerationResult"
                    },
                    "tagAssignment": {
                      "$ref": "#/components/schemas/TagAssignmentStatus"
                    },
                    "tagAssignmentErrorStatus": {
                      "$ref": "#/components/schemas/TagAssignmentErrorStatus"
                    }
                  },
                  "required": [
                    "question",
                    "moderation",
                    "tagAssignment"
                  ]
                }
              }
            }
          },
          "201": {
            "description": "作成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "question": {
                      "$ref": "#/components/schemas/Question"
                    },
                    "moderation": {
                      "$ref": "#/components/schemas/ModerationResult"
                    },
                    "tagAssignment": {
                      "$ref": "#/components/schemas/TagAssignmentStatus"
                    },
                    "tagAssignmentErrorStatus": {
                      "$ref": "#/components/schemas/TagAssignmentErrorStatus"
                    }
                  },
                  "required": [
                    "question",
                    "moderation",
                    "tagAssignment"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "不正なタグ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "選択されたタグが見つかりません"
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "409": {
            "description": "同じキーで異なる内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "同じ投稿操作に異なる内容が指定されています"
                }
              }
            }
          },
          "500": {
            "description": "作成失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "投稿の作成に失敗しました"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions/{id}": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "質問詳細を取得",
        "security": [
          {
            "supabaseSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "post-1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "質問詳細",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "質問が見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Not found"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions/{id}/resolve": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "質問の募集を終了",
        "security": [
          {
            "supabaseSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "post-1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "募集終了",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "question": {
                      "$ref": "#/components/schemas/Question"
                    }
                  },
                  "required": [
                    "question"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "質問者以外",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "質問が見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Not found"
                }
              }
            }
          },
          "409": {
            "description": "操作不可",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "操作できない質問の状態です"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/question-tags": {
      "get": {
        "tags": [
          "question-tags"
        ],
        "summary": "Q&Aで使用するタグ候補を取得",
        "security": [
          {
            "supabaseSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Q&Aタグ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/QuestionTagCategory"
                      }
                    }
                  },
                  "required": [
                    "categories"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "未認証",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}